Save quantities computed at runtime
