+++ src/basta/utils_general.py
@@ -14,7 +14,7 @@ import h5py  # type: ignore[import]
 import numpy as np
 import scipy.linalg  # type: ignore[import]
 
-from basta import core, constants, distances, errors, freq_fit, glitch_fit
+from basta import core, constants, distances, errors, freq_fit, glitch_fit, utils_priors
 from basta import fileio as fio
 from basta import utils_seismic as su
 from basta.__about__ import __version__
@@ -1146,127 +1146,6 @@ def get_epsilondifferences(
     return epsilondiff_dict
 
 
-def get_limits(
-    inputstar: core.InputStar,
-    inferencesettings: core.InferenceSettings,
-    distancelimits: dict[str, tuple[float, float]] | None = None,
-):
-    """
-    They are interpreted as
-    - `min`: an absolute lower bound.
-    - `max`: an absolute upper bound.
-    - `abstol`: symmetric bound around the observed stellar property
-    - `sigmacut`: symmetric bound around the observed stellar property
-
-    As they can overlap, we intersect all apllicable constraints.
-    """
-    limits: dict[str, tuple[float, float]] = {}
-    priors = inferencesettings.boxpriors
-    if distancelimits is None:
-        distancelimits = {}
-
-    params = {
-        **inputstar.classicalparams.params,
-        **inputstar.globalseismicparams.params,
-    }
-
-    # Unpack special cases such as gridcut and dnufrac
-    gridcut = priors.get("gridcut")
-    gridcut_limits: dict[str, tuple[float, float]] = (
-        gridcut if isinstance(gridcut, dict) else {}
-    )
-
-    dnufrac = priors.get("dnufrac")
-    dnufrac_limits: dict[str, tuple[float, float]] = {}
-    if isinstance(dnufrac, core.PriorEntry):
-        dnutype = "dnufit"
-        if dnutype in params:
-            dnu_value, dnu_error = inputstar.globalseismicparams.get_scaled(dnutype)
-            frac = dnufrac.kwargs[dnutype]
-            three_sigma = 3 * dnu_error
-            delta = min(three_sigma, frac * dnu_value)
-            dnufrac_limits = {dnutype: (max(0.0, dnu_value - delta), dnu_value + delta)}
-
-    all_dimensions = (
-        set(priors.keys())
-        | set(distancelimits.keys())
-        | set(gridcut_limits.keys())
-        | set(dnufrac_limits.keys())
-    )
-
-    for dimension in all_dimensions:
-        if dimension in [
-            "gridcut",
-            "dnufrac",
-        ]:
-            continue  # as we are processing its contents instead.
-
-        star_param = params.get(dimension)
-        prior_entry = priors.get(dimension)
-        dist_limit = distancelimits.get(dimension)
-        gridcut_limit = gridcut_limits.get(dimension)
-        dnufrac_limit = dnufrac_limits.get(dimension)
-
-        prior_bounds: list[tuple[float, float]] = []
-        if prior_entry:
-            # star_param must be Fitparam, since we have priors.
-            assert isinstance(star_param, tuple)  # Fitparam is a tuple
-            kwargs = prior_entry.kwargs
-            if "min" in kwargs:
-                prior_bounds.append((kwargs["min"], float("inf")))
-            if "max" in kwargs:
-                prior_bounds.append((float("-inf"), kwargs["max"]))
-            if "abstol" in kwargs:
-                tol = kwargs["abstol"]
-                if star_param is None:
-                    raise ValueError(
-                        f"Missing stellar value for dimension '{dimension}'"
-                    )
-                prior_bounds.append((star_param[0] - tol, star_param[0] + tol))
-            if "sigmacut" in kwargs:
-                if star_param is None:
-                    raise ValueError(
-                        f"Missing stellar value for dimension '{dimension}'"
-                    )
-                sigma = star_param[1]
-                sigcut = kwargs["sigmacut"]
-                prior_bounds.append(
-                    (star_param[0] - sigcut * sigma, star_param[0] + sigcut * sigma)
-                )
-
-        prior_limit = None
-        if prior_bounds:
-            lower = max(b[0] for b in prior_bounds)
-            upper = min(b[1] for b in prior_bounds)
-            if lower > upper:
-                raise ValueError(
-                    f"Incompatible prior limits for '{dimension}': {prior_bounds}"
-                )
-            prior_limit = (lower, upper)
-
-        candidate_limits = [
-            l
-            for l in [prior_limit, dist_limit, gridcut_limit, dnufrac_limit]
-            if l is not None
-        ]
-
-        # Combine limits by intersecting ranges if this applies
-        if candidate_limits:
-            lower = max(l[0] for l in candidate_limits)
-            upper = min(l[1] for l in candidate_limits)
-            if lower > upper:
-                raise ValueError(
-                    f"Incompatible combined limits for '{dimension}': {candidate_limits}"
-                )
-            limits[dimension] = (lower, upper)
-        else:
-            raise NotImplementedError(
-                f"Limit generation not implemented for '{dimension}'"
-            )
-
-    return limits
-
-
 def setup_star(
     inputstar: core.InputStar,
     inferencesettings: core.InferenceSettings,
@@ -1342,7 +1221,7 @@ def setup_star(
         )
 
     # Translate boxpriors into ranges, depending on the given star
-    limits = get_limits(
+    limits = utils_priors.get_limits(
         inputstar=inputstar,
         inferencesettings=inferencesettings,
         distancelimits=distancelimits,
